requirements.txt
salesforce-lavis
ftfy
regex
datasets==2.12.0
icrawler==0.6.6
iopath==0.1.10
jsonlines==3.1.0
lmdb==1.4.1
networkx==2.5.1
nltk==3.8.1
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
PyYAML==6.0
Requests==2.31.0
rouge_score==0.1.2
six==1.16.0
spacy==3.5.3
tqdm==4.65.0
timm==0.9.2
--extra-index-url https://download.pytorch.org/whl/cu102
torch==1.12.1+cu102
torch_scatter==2.1.1
torchvision==0.13.1+cu102
transformers==4.34.0
